refactored Google Cloud BigQuery dependent code to fit.py; the rest remains in main.py. Added Postman tests for both.
