+++ app.yaml
@@ -1,6 +1,7 @@
 runtime: python
 env: flex
-entrypoint: gunicorn --timeout 3600 --workers 4 --threads 12 -b :$PORT main:app
+# --threads 48 fails with ERROR: (gcloud.app.deploy) Error Response: [13] An internal error occurred during deployment. You may need to delete this version manually.
+entrypoint: gunicorn --timeout 3600 --workers 8 --threads 10 -b :$PORT fit:app
 
 runtime_config:
   python_version: 2
